refactor: ♻️ (cms, model) パスエイリアスの変更に合わせて、モジュールのパス指定を修正した。
