+++ packages/cms/src/converter/article.converter.ts
@@ -2,7 +2,7 @@ import type { Article, ArticleDetail } from '@nitic-astronomy/model';
 import { convertImage } from './image.converter';
 import { convertMember } from './member.converter';
 import { convertTag } from './tag.converter';
-import type { MicroCmsArticle } from '~cms/infra/microcms.type';
+import type { MicroCmsArticle } from '#cms/infra/microcms.type';
 
 export const convertArticle = (article: MicroCmsArticle): Article => ({
   id: article.id,
