+++ packages/cms/src/converter/tag.converter.ts
@@ -1,5 +1,5 @@
 import type { Tag } from '@nitic-astronomy/model';
-import type { MicroCmsTag } from '~cms/infra/microcms.type';
+import type { MicroCmsTag } from '#cms/infra/microcms.type';
 
 export const convertTag = (tag: MicroCmsTag): Tag => ({
   id: tag.id,
