+++ packages/cms/src/index.ts
@@ -1,16 +1,16 @@
-export { getArticles } from '~cms/service/article/get-articles.service';
-export type { GetArticlesQuery, GetArticlesResponse, GetArticles } from '~cms/service/article/get-articles.service';
-export { getArticleIds } from '~cms/service/article/get-article-ids.service';
-export type { GetArticleIdsResponse, GetArticleIds } from '~cms/service/article/get-article-ids.service';
-export { getArticleDetail } from '~cms/service/article/get-article-detail.service';
-export type { GetArticleDetailQuery, GetArticleDetailResponse, GetArticleDetail } from '~cms/service/article/get-article-detail.service';
+export { getArticles } from '#cms/service/article/get-articles.service';
+export type { GetArticlesQuery, GetArticlesResponse, GetArticles } from '#cms/service/article/get-articles.service';
+export { getArticleIds } from '#cms/service/article/get-article-ids.service';
+export type { GetArticleIdsResponse, GetArticleIds } from '#cms/service/article/get-article-ids.service';
+export { getArticleDetail } from '#cms/service/article/get-article-detail.service';
+export type { GetArticleDetailQuery, GetArticleDetailResponse, GetArticleDetail } from '#cms/service/article/get-article-detail.service';
 
-export { getMembers } from '~cms/service/member/get-members.service';
-export type { GetMembersQuery, GetMembersResponse, GetMembers } from '~cms/service/member/get-members.service';
+export { getMembers } from '#cms/service/member/get-members.service';
+export type { GetMembersQuery, GetMembersResponse, GetMembers } from '#cms/service/member/get-members.service';
 
-export { getTags } from '~cms/service/tag/get-tags.service';
-export type { GetTagsQuery, GetTagsResponse, GetTags } from '~cms/service/tag/get-tags.service';
-export { getTagIds } from '~cms/service/tag/get-tag-ids.service';
-export type { GetTagIdsResponse, GetTagIds } from '~cms/service/tag/get-tag-ids.service';
-export { getTagDetail } from '~cms/service/tag/get-tag-detail.service';
-export type { GetTagDetailQuery, GetTagDetailResponse, GetTagDetail } from '~cms/service/tag/get-tag-detail.service';
+export { getTags } from '#cms/service/tag/get-tags.service';
+export type { GetTagsQuery, GetTagsResponse, GetTags } from '#cms/service/tag/get-tags.service';
+export { getTagIds } from '#cms/service/tag/get-tag-ids.service';
+export type { GetTagIdsResponse, GetTagIds } from '#cms/service/tag/get-tag-ids.service';
+export { getTagDetail } from '#cms/service/tag/get-tag-detail.service';
+export type { GetTagDetailQuery, GetTagDetailResponse, GetTagDetail } from '#cms/service/tag/get-tag-detail.service';
