+++ packages/cms/src/service/article/get-articles.service.ts
@@ -1,8 +1,8 @@
 import type { Article, Tag } from '@nitic-astronomy/model';
-import { convertArticle } from '~cms/converter/article.converter';
-import { client } from '~cms/infra/client.factory';
-import type { MicroCmsArticle } from '~cms/infra/microcms.type';
-import { MicroCmsApiEndpoint } from '~cms/infra/microcms.type';
+import { convertArticle } from '#cms/converter/article.converter';
+import { client } from '#cms/infra/client.factory';
+import type { MicroCmsArticle } from '#cms/infra/microcms.type';
+import { MicroCmsApiEndpoint } from '#cms/infra/microcms.type';
 
 export type GetArticlesQuery = {
   limit?: number;
