+++ packages/model/src/index.ts
@@ -1,11 +1,11 @@
-export type { Article, ArticleDetail } from '~model/model/article.model';
-export type { Member } from '~model/model/member.model';
-export type { Tag } from '~model/model/tag.model';
-export type { Image } from '~model/model/image.model';
-export type { Id } from '~model/model/id.model';
+export type { Article, ArticleDetail } from '#model/model/article.model';
+export type { Member } from '#model/model/member.model';
+export type { Tag } from '#model/model/tag.model';
+export type { Image } from '#model/model/image.model';
+export type { Id } from '#model/model/id.model';
 
-export { ArticleSchema, ArticleDetailSchema } from '~model/model/article.model';
-export { MemberSchema } from '~model/model/member.model';
-export { TagSchema } from '~model/model/tag.model';
-export { ImageSchema } from '~model/model/image.model';
-export { IdSchema } from '~model/model/id.model';
+export { ArticleSchema, ArticleDetailSchema } from '#model/model/article.model';
+export { MemberSchema } from '#model/model/member.model';
+export { TagSchema } from '#model/model/tag.model';
+export { ImageSchema } from '#model/model/image.model';
+export { IdSchema } from '#model/model/id.model';
